Move universal constants to separate module
