+++ cluster-autoscaler/core/scale_down_test.go
@@ -38,6 +38,7 @@ import (
 	kube_util "k8s.io/autoscaler/cluster-autoscaler/utils/kubernetes"
 	scheduler_util "k8s.io/autoscaler/cluster-autoscaler/utils/scheduler"
 	. "k8s.io/autoscaler/cluster-autoscaler/utils/test"
+	"k8s.io/autoscaler/cluster-autoscaler/utils/units"
 	"k8s.io/client-go/kubernetes/fake"
 	core "k8s.io/client-go/testing"
 
@@ -863,7 +864,7 @@ var defaultScaleDownOptions = context.AutoscalingOptions{
 	MinCoresTotal:                 0,
 	MinMemoryTotal:                0,
 	MaxCoresTotal:                 config.DefaultMaxClusterCores,
-	MaxMemoryTotal:                config.DefaultMaxClusterMemory * config.Gigabyte,
+	MaxMemoryTotal:                config.DefaultMaxClusterMemory * units.Gigabyte,
 }
 
 func TestScaleDownEmptyMultipleNodeGroups(t *testing.T) {
