+++ cluster-autoscaler/main.go
@@ -40,6 +40,7 @@ import (
 	"k8s.io/autoscaler/cluster-autoscaler/simulator"
 	"k8s.io/autoscaler/cluster-autoscaler/utils/errors"
 	kube_util "k8s.io/autoscaler/cluster-autoscaler/utils/kubernetes"
+	"k8s.io/autoscaler/cluster-autoscaler/utils/units"
 	kube_client "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -144,8 +145,8 @@ func createAutoscalingOptions() context.AutoscalingOptions {
 		glog.Fatalf("Failed to parse flags: %v", err)
 	}
 	// Convert memory limits to bytes.
-	minMemoryTotal = minMemoryTotal * config.Gigabyte
-	maxMemoryTotal = maxMemoryTotal * config.Gigabyte
+	minMemoryTotal = minMemoryTotal * units.Gigabyte
+	maxMemoryTotal = maxMemoryTotal * units.Gigabyte
 
 	return context.AutoscalingOptions{
 		CloudConfig:                      *cloudConfig,
